更改右滑关闭的依赖库，解决8.0以上系统不兼容右滑的问题；修改视频解析逻辑;解决视频播放错乱的问题,更换了最新版的饺子播放器；整理项目结构，包的路径等

+++ app/src/main/java/com/chaychan/news/model/event/DetailCloseEvent.java
@@ -9,7 +9,7 @@ public class DetailCloseEvent {
 
     private String channelCode;
     private int position;
-    private int progress;
+    private long progress;
     private int commentCount;
 
     public String getChannelCode() {
@@ -28,11 +28,11 @@ public class DetailCloseEvent {
         this.position = position;
     }
 
-    public int getProgress() {
+    public long getProgress() {
         return progress;
     }
 
-    public void setProgress(int progress) {
+    public void setProgress(long progress) {
         this.progress = progress;
     }
